requirements.txt

fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
google-genai
